MAJOR REFACTOR, renamed net.pixlies to net.pixlies.core because of conflict with nations
