+++ src/main/java/net/pixlies/core/moderation/PunishmentType.java
@@ -1,0 +1,9 @@
+package net.pixlies.core.moderation;
+
+public enum PunishmentType {
+
+    BAN,
+    MUTE,
+    WARN
+
+}
