+++ src/main/java/net/pixlies/core/utils/CC.java
@@ -1,0 +1,11 @@
+package net.pixlies.core.utils;
+
+import org.bukkit.ChatColor;
+
+public final class CC {
+
+    public static String format(String message) {
+        return ChatColor.translateAlternateColorCodes('&', message);
+    }
+
+}
