--- src/main/java/net/pixlies/moderation/PunishmentType.java
@@ -1,9 +1,0 @@
-package net.pixlies.moderation;
-
-public enum PunishmentType {
-
-    BAN,
-    MUTE,
-    WARN
-
-}
