--- src/main/java/net/pixlies/utils/Palette.java
@@ -1,17 +1,0 @@
-package net.pixlies.utils;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import org.bukkit.ChatColor;
-
-@Data
-@AllArgsConstructor
-public class Palette {
-
-    public static final Palette GREEN_THICK = new Palette(ChatColor.BOLD, ChatColor.GREEN, ChatColor.DARK_GREEN);
-
-    private ChatColor effect;
-    private ChatColor primary;
-    private ChatColor accent;
-
-}
